requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
catboost==1.2.2
shap==0.43.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
kafka-python==2.0.2
redis==5.0.1
mlflow==2.9.2
prometheus-client==0.19.0
evidently==0.4.0
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
fastapi-cache2==0.2.1
python-json-logger==2.0.7 